chore: check if template files already exist in fs

+++ extension/src/utils/buildFolderStructure.ts
@@ -1,5 +1,6 @@
 import * as vscode from 'vscode'
 import { TemplateFolderStructure } from '../types'
+import * as fs from 'fs'
 
 const createFileInVSCode = async (fileItem: TemplateFolderStructure) => {
   if (!fileItem) {
@@ -12,7 +13,17 @@ const createFileInVSCode = async (fileItem: TemplateFolderStructure) => {
     return await buildFolderStructure(content)
   }
 
-  // TODO - Check if file already exists at path
+  const doesFileAlreadyExist = fs.existsSync(path)
+  const formattedPath = path.replace(/\\/g, '/')
+
+  if (doesFileAlreadyExist) {
+    const choices = ['Overwrite', 'Keep Existing']
+
+    const choice = await vscode.window.showInformationMessage(`File ${formattedPath} already exists`, ...choices)
+
+    if (choice === choices[1]) return
+  }
+
   await vscode.workspace.fs.writeFile(vscode.Uri.file(path), new TextEncoder().encode(content))
 }
